libpcre2: Default enable JIT support for aarch64
JIT support in pcre2 allows for extra performance for regex operations in applications that support it.
As outlined in https://pcre.org/current/doc/html/pcre2jit.html#SEC2 64-bit ARM is supported.

I tested this on an GL.Inet MT6000 which is an aarch64 device and to my knowledge everything works as expected.
The primary application I tested this on was haproxy, which makes use pcre for several operations.

If there are no known downsides or known breakages I suggest to default-enable this feature for aarch64.

Signed-off-by: Christian Lachner <gladiac@gmail.com>